Bugfix for NPE on dimension load with other chunkloaders
The getPlayerOrganiser method references the playerOrganisers hashmap without checking to see if it is null. It will be null if the activate() method of TileChunkLoader base is called by another chunkloading mod before the chunk manager is loaded. A simple null check won't work in this case, so I checked the load flag and called the load() method to initial the chunk manager if it hasn't been loaded yet.